+++ crates/sindri-assets/src/decode/mod.rs
@@ -6,6 +6,7 @@
 
 mod font;
 mod prefab;
+mod profile;
 mod scene;
 mod sheet;
 mod text;
@@ -16,6 +17,7 @@ mod tests;
 
 pub use font::{FontAsset, FontAssetDecoder};
 pub use prefab::PrefabAssetDecoder;
+pub use profile::ProfileAssetDecoder;
 pub use scene::SceneAssetDecoder;
 pub use sheet::SpriteSheetAssetDecoder;
 pub use text::TextAssetDecoder;